move ToMemcachedValue from parser/mod to value_serializer.rs and use sealed trait pattern to prevent external impl on consumer types
